add CI build using GH Actions
this runs for all PRs and builds the project against python 3.10 (latest
packaged version in Ubuntu 22.04 (current LTS e.g. shipped in WSL2) and
python 3.11 (latest stable release currently available) on the different
platforms.

as there are no tests yet it doesn't trigger any tests. this will be
added if/when tests are being added.