refactor app launch
based on this refactoring it'll be possible to add other launchers (e.g.
with a continuous video stream to be processed) in the future.

this moves the application launch into an own module and gives it a
`__main__.py` which allows running it using
`python3 -m single_image_proc`. if you're not in a venv you'll have to
prepend the call with `poetry run` to run it in the venv.