only print new/changed patterns in log
so far during every processing step all patterns which had been found
have been printed into the log. this is spamming the log and making it
impossible to see, which patterns are new and which ones aren't.

to compare two patterns the `__eq__` on `Pattern` has been implemented.
in the process also `__str__` has been implemented to aid with
debugging.

closes #17