remove author & copyright from Klims in `app.py`
this was automatically generated by his IDE when he created the app,
however this is not appropriate anymore as the file has been heavily
edited (the files which remain largely unchanged in turn don't have a
copyright & author notice anyway).

Klims by email confirmed that the files can be licensed as GPLv3 and the
commit introducing them was done with him as author, thus authorship is
still established. as the code was provided by him at the FHGR it is
also unrelated to CEDES (his regular employer).